Increase agent count to 5000
This results in a ~10x increase in overall traffic.

This is enough to cause observable traffic congestion effects.
When the resulting simulation is visualized in Via, a significant amount
of agents are slowed to ~50% of the free speed at the 17:54 mark.
This is observable both in the mixed and segregated scenarios.

Note: Since the free version of Via only supports <= 500 agents, the
resulting visualization displays only a small subset of the entire
population. Nevertheless, it is enough to observe the impact of
population on congestion.